Fix memory leak by adding tiny_queue_destroy
You need to deallocate the queue when you have finished to use it; if
you don't it you have a memory leak. So I've added a new function,
'tiny_queue_destroy', that deallocate all elements of the queue and the
queue structure itself.

Now the examples 'hello.c' and 'hello_thread.c' works correcly, while
'hello_struct.c' needs other coding, because you can't use free() to
deallocate recursively a structure (so the memory leak on this code
still remains).